Fix errors preventing compilation with clang
Clang was the first to compile; GCC is not happy with the current Makefiles
somehow.

Clang complained on macro errors and a missing return value.  Strange that MSVC
would let the return slip.  The macros errors are curious as well: the
expressions seem valid, and the fix don't seem to change their semantics, but
the compiler is happy now.