M68k: Fix incorrect address calculation in debugger
Branch instructions compute the decoded effective address using
`instruction_address` and `instruction_register` from the context, but they were
not overridden.  This was causing any lines beyond the current one to display
incorrect operands and computed jump addresses in the disassembly.

The correct solution would be to /clone/ the full state of the M68k in
m68k_decode, to be sure that we do not change it in anyway by running it on
future instructions.  Unfortunately, that would require populating the opcode
table for each clone (since instructions are bound to one context at creation),
and that is prohibitive.