Limit sprites per scanline to avoid infinite loop
Quackshot runs into an infinite loop otherwise.  Maybe the sprite linked list is
malformed, or there is an escape condition we fail to emulate.

There is a hardware limit, but using the commented condition makes sprites
disappear in Sonic 1.

Using a higher limit does not impact Sonic 1 and let Quackshot run past the
intro.