Fix JSON_SET/JSON_GET macros identifier concatenation
Neither gcc nor clang would have it.  They complained that concatenating `->`
and anything resulted in an invalid identifier.  The way I understand it, `->`
is an operator, not part of an identifier, so you actually don't need to resort
to concatenation using `##` here.

Using `s->name` is enough for the replacement to take place.