Revamp Makefile and run.sh
No more `make clean` when headers change!  Now, every dependencies for C files
are generated with the -MMD flag and we recompile only what's needed.

Also, the duplication of debug/release builds has been moved to the run.sh
script, since that was dirty in the Makefile.  As a side-effect, for cleaning
debug/release builds you need to run `./run.sh clean`.