Separate ROM and RAM, unmute invalid I/O
We allocated 0x1000000 bytes to cover the whole memory map of the Genesis but in practice, its central chunk mostly consists of memory-mapped I/O devices. This wastes some memory, but the real issue here is that invalid read/writes were silenced. Now, we are in a better position to debug bad addressing, missing mirroring, etc.