Fix a couple of GCC warnings
operands.c: address_registers and data_registers are not of the same sign, so
returning one of them is awkward.  But, since we always seem to cast the return
value as uint32_t anyway, might as well convert the data_register directly.

renderer.c: pointer-to-int conversion.  The fix is a bit of a mouthful... not
sure it's the canonical way either, but it works.