psg: Fix frequency divisor (it sounds good now!)
The PSG divides the master clock by 16, and we previously clocked directly at
the master frequency.

Went back using doubles for PSG_CLOCKS_PER_SAMPLE.  The current value is
5.07305130385, so rounded down to 5 the difference is not huge, but the pitches
sound better with the extra precision.  (Besides, it /is/ the correct thing to
do).

Now Sonic 1 Stage 1 sounds okay!  The PSG is used for accompanying notes though,
not the main melody: that's the YM2612.