settings: Add default values when loading from an existing file
We had default values but only if the file did not exist.  If the settings file
existed, but we didn't have any key for a specific setting, that setting would
default to zero.

Now, when loading the file, all settings have a default value that is used if no
key for that setting is present in the file.